% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxdbHandling.R
\name{basicDecoder}
\alias{basicDecoder}
\title{convert binary output of GET()$content to list}
\usage{
basicDecoder(x)
}
\arguments{
\item{x}{string suitable for input to GET as GET(x)}
}
\value{
output of fromJSON, typically a list
}
\description{
convert binary output of GET()$content to list
}
\examples{
cl = basicDecoder("https://api.pharmacodb.com/v1/cell_lines")
unlist(cl)
}
