% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compounds_v1}
\alias{compounds_v1}
\alias{tissues_v1}
\alias{cell_lines_v1}
\alias{datasets_v1}
\alias{CCLE_drts}
\title{compounds_v1: serialization of compounds info from PharmacoDb v1}
\format{
S4Vectors DataFrame instance

S4Vectors DataFrame instance

S4Vectors DataFrame instance

S4Vectors DataFrame instance

DRTraceSet instance
}
\source{
PharmacoDb Sept 2017

PharmacoDb Sept 2017

PharmacoDb Sept 2017

PharmacoDb Sept 2017

PharmacoDb April 2018
}
\usage{
compounds_v1

tissues_v1

cell_lines_v1

datasets_v1

CCLE_drts
}
\description{
compounds_v1: serialization of compounds info from PharmacoDb v1
}
\examples{
data(compounds_v1)
head(compounds_v1)
data(tissues_v1)
head(tissues_v1)
data(cell_lines_v1)
head(cell_lines_v1)
data(datasets_v1)
head(datasets_v1)
data(CCLE_drts)
CCLE_drts
}
\keyword{datasets}
