% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shin.R
\name{compoundsByCell}
\alias{compoundsByCell}
\title{initial version of compound browser over pharmacoDb cells}
\usage{
compoundsByCell()
}
\value{
only used for side effect of running shiny app
}
\description{
initial version of compound browser over pharmacoDb cells
}
\note{
Simple shiny app demonstrating coverage of PharmacoDb 
compounds by CHEBI.  If a cell line selected is not present 
in selected dataset, the app will wait for a 
compatible selection to be made.
}
\examples{
if (!requireNamespace("shiny")) stop("install shiny to use compoundsByCell")
if (interactive()) print(compoundsByCell())
}
