% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{iriCCLE}
\alias{iriCCLE}
\title{obtain an example trace set stored locally, for irinotecan and selected cell lines}
\usage{
iriCCLE()
}
\description{
obtain an example trace set stored locally, for irinotecan and selected cell lines
}
\examples{
iri = iriCCLE()
iri
plot(iri)
}
