% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxdbHandling.R
\name{rxdbQuery_v1}
\alias{rxdbQuery_v1}
\title{very simple query formulation, build queries using endpoints of bhklab PharmacoDB API}
\usage{
rxdbQuery_v1(..., url = "https://api.pharmacodb.com/v1/",
  decoder = basicDecoder)
}
\arguments{
\item{\dots}{typically a string representing an API endpoint, will be processed by unlist() and then to paste0 preceded by \code{url}}

\item{url}{of a PharmacoDB server API target}

\item{decoder}{a function of one argument that will be applied to API response (typically JSON)}
}
\description{
very simple query formulation, build queries using endpoints of bhklab PharmacoDB API
}
\examples{
qout = rxdbQuery_v1("cell_lines") # yields 30; append "?all=true" to retrieve all
sapply(qout, function(x) x[[2]])
}
