% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxdbHandling.R
\name{topEndpoints_v1}
\alias{topEndpoints_v1}
\title{enumerate top level endpoint terms for bhklab PharmacoDB API}
\usage{
topEndpoints_v1()
}
\value{
a character vector of available endpoints
}
\description{
enumerate top level endpoint terms for bhklab PharmacoDB API
}
\examples{
topEndpoints_v1()
}
