% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shin.R
\name{compoundsByCell}
\alias{compoundsByCell}
\title{initial version of compound browser over pharmacoDb cells}
\usage{
compoundsByCell()
}
\description{
initial version of compound browser over pharmacoDb cells
}
\note{
simple shiny app demonstrating coverage of pharmacoDb compounds by CHEBI
}
