% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkcomb.R
\docType{class}
\name{DRProfile-class}
\alias{DRProfile-class}
\alias{getDrugs}
\alias{getDrugs,DRProfSet-method}
\title{DRProfSet is a class for managing dose-response information about cell lines from a pharmacogenomics dataset}
\usage{
getDrugs(x)
}
\arguments{
\item{x}{instance of DRProfSet}
}
\value{
getDrugs: character vector
}
\description{
DRProfSet is a class for managing dose-response information about cell lines from a pharmacogenomics dataset

getDrugs extracts drug list
}
